refactor: split transformers and sam integrations
