+++ rankseg/README.md
@@ -7,10 +7,13 @@ This directory contains the Python package modules behind the public `rankseg` A
 - `_rankseg.py`: defines the `RankSEG` class and the main prediction entry point
 - `_rankseg_algo.py`: lower-level solver implementations used by `RankSEG`
 - `distribution.py`: probability-distribution utilities used by the algorithms
-- `transformers.py`: Hugging Face Transformers compatibility helpers. Use `postprocess` for inference; use the restore helpers only when probability maps are needed directly.
+- `integration/`: compatibility adapters for external model APIs
+  - `transformers.py`: standard Hugging Face semantic segmentation helpers. Use `postprocess` for inference; use `restore_semantic_probs` when probability maps are needed directly.
+  - `sam.py`: SAM-family adapters. Use `Sam1`, `Sam2`, or `Sam3` for family-specific mask restoration and postprocessing.
 
 ## Import path
 
 ```python
-from rankseg.transformers import postprocess, restore_semantic_probs, restore_sam_mask_probs
+from rankseg.integration.transformers import postprocess, restore_semantic_probs
+from rankseg.integration.sam import Sam1, Sam2, Sam3
 ```
